# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_requestRegistrationOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestRegistrationOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Phone already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Register and request a verification OTP via SMS",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register/resend-otp": {
      "post": {
        "operationId": "AuthController_resendRegistrationOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResendRegistrationOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP resent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "No pending registration found for this phone",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Resend registration OTP to an existing pending registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/register/verify": {
      "post": {
        "operationId": "AuthController_verifyRegistrationOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyRegistrationOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered; returns access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Verify OTP and complete registration",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Login with phone and password",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/request-otp": {
      "post": {
        "operationId": "AuthController_requestPasswordResetOtp",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestPasswordResetOtpDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OTP sent (or silently ignored if phone not found)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OtpSentResponseDto"
                }
              }
            }
          }
        },
        "summary": "Request a password-reset OTP via SMS",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Password updated; returns new tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired OTP",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Reset password using OTP",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Returns new access + refresh tokens",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refresh access token using a refresh token",
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LogoutDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Logged out successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LogoutResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Revoke refresh token and logout",
        "tags": [
          "Auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Market Place API",
    "description": "Online marketplace REST API — user registration, product listings, real-time chat, ratings, and admin moderation.\n\n> **WebSocket (Chat):** Real-time messaging uses a Socket.io gateway at the `/chat` namespace. That interface is not documented here; see the project README for event names and payloads.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RequestRegistrationOtpDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Full name (2–150 chars)",
            "example": "Ahmed Ali",
            "minLength": 2,
            "maxLength": 150
          },
          "phone": {
            "type": "string",
            "description": "Phone number (E.164 format)",
            "example": "+201234567890"
          },
          "password": {
            "type": "string",
            "description": "Password — must contain letters and numbers (8–64 chars)",
            "example": "Secret123",
            "minLength": 8,
            "maxLength": 64
          }
        },
        "required": [
          "name",
          "phone",
          "password"
        ]
      },
      "OtpSentDataDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "OTP sent"
          },
          "otp": {
            "type": "string",
            "example": "000000",
            "description": "Only present when OTP_DEV_MODE=true (fixed to 000000 for console provider)"
          }
        },
        "required": [
          "message"
        ]
      },
      "OtpSentResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/OtpSentDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "ErrorDetailDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "example": 409
          },
          "message": {
            "type": "string",
            "example": "Conflict"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-03-28T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/auth/register/request-otp"
          }
        },
        "required": [
          "code",
          "message",
          "timestamp",
          "path"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "number",
            "example": 409
          },
          "data": {
            "type": "object",
            "additionalProperties": false,
            "example": null,
            "nullable": true
          },
          "error": {
            "$ref": "#/components/schemas/ErrorDetailDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data",
          "error"
        ]
      },
      "ResendRegistrationOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Phone number (E.164 format)",
            "example": "+201234567890"
          }
        },
        "required": [
          "phone"
        ]
      },
      "VerifyRegistrationOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Phone number (E.164 format)",
            "example": "+201234567890"
          },
          "otp": {
            "type": "string",
            "description": "One-time password sent via SMS (4–8 digits)",
            "example": "000000",
            "minLength": 4,
            "maxLength": 8
          },
          "transactionReqID": {
            "type": "string",
            "description": "Akedly transaction request ID",
            "example": "68b4a1e8d686446a498008bd"
          }
        },
        "required": [
          "phone",
          "otp"
        ]
      },
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "phone": {
            "type": "string",
            "example": "+201012345678"
          }
        },
        "required": [
          "id",
          "phone"
        ]
      },
      "TokenDataDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/AuthUserDto"
          },
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          },
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken",
          "refreshToken"
        ]
      },
      "TokenResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/TokenDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "User phone number (E.164 format)",
            "example": "+201234567890"
          },
          "password": {
            "type": "string",
            "description": "User password (8–64 chars)",
            "example": "Secret123",
            "minLength": 8,
            "maxLength": 64
          }
        },
        "required": [
          "phone",
          "password"
        ]
      },
      "RequestPasswordResetOtpDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Registered phone number to receive the reset OTP",
            "example": "+201234567890"
          }
        },
        "required": [
          "phone"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "description": "Phone number associated with the account",
            "example": "+201234567890"
          },
          "otp": {
            "type": "string",
            "description": "One-time password received via SMS (4–8 digits)",
            "example": "000000",
            "minLength": 4,
            "maxLength": 8
          },
          "transactionReqID": {
            "type": "string",
            "description": "Akedly transaction request ID",
            "example": "68b4a1e8d686446a498008bd"
          },
          "newPassword": {
            "type": "string",
            "description": "New password — must contain letters and numbers (8–64 chars)",
            "example": "NewSecret123",
            "minLength": 8,
            "maxLength": 64
          },
          "confirmPassword": {
            "type": "string",
            "description": "Must match newPassword",
            "example": "NewSecret123",
            "minLength": 8,
            "maxLength": 64
          }
        },
        "required": [
          "phone",
          "otp",
          "newPassword",
          "confirmPassword"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "JWT refresh token",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "The refresh token to revoke",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "LogoutDataDto": {
        "type": "object",
        "properties": {}
      },
      "LogoutResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "number",
            "example": 200
          },
          "data": {
            "$ref": "#/components/schemas/LogoutDataDto"
          }
        },
        "required": [
          "success",
          "statusCode",
          "data"
        ]
      },
...
    }
  }
}